refactor: extract SVG timer to separate CircularProgressTimer component
- Create dedicated CircularProgressTimer component for better maintainability
- Extract 150+ lines of SVG logic into reusable component
- Improve code organization and separation of concerns
- Enable easier testing and potential reuse in other parts of the app
- Keep all timer-specific logic and styling encapsulated
